Development

+++ src/app/products/admin/api/route.js
@@ -1,9 +1,11 @@
 import { NextResponse } from 'next/server';
 import { collection, getDocs, addDoc, updateDoc, deleteDoc, doc } from 'firebase/firestore';
-import { db } from '../../../../lib/firebase/config.js';
+import { useAuthState } from 'react-firebase-hooks/auth';
+import { auth, db } from '../../../../lib/firebase/config.js';
 
 export async function GET() {
 	try {
+		console.log('getting prods');
 		const querySnapshot = await getDocs(collection(db, 'products'));
 		const products = [];
 		querySnapshot.forEach((doc) => {
@@ -40,13 +42,29 @@ export async function PUT(request) {
 }
 
 export async function DELETE(request) {
+	console.log('asd');
+	console.log(request);
+
 	try {
-		const { id } = await request.json();
+		// const [user] = useAuthState(auth);
+		console.log('asd2');
+		// if (!user) {
+		// 	return NextResponse.json({ error: 'User not authenticated' }, { status: 401 });
+		// }
+		const { productId } = await request.body;
+		console.log('Product id: ' + productId);
+		console.log('Request body: ' + request.body);
+		console.log('Request body str: ' + JSON.stringify(request.body));
+
+		if (!productId) {
+			return NextResponse.json({ error: 'Product ID is required' }, { status: 404 });
+		}
+		const { id } = request.json();
 		const docRef = doc(db, 'products', id);
 		await deleteDoc(docRef);
 		return NextResponse.json({ id });
 	} catch (error) {
-		console.error('Error deleting product:', error);
-		return NextResponse.error();
+		console.log(error);
+		return NextResponse.json({ error: 'Product ID is required. + ' + error }, { status: 404 });
 	}
 }
